+++ app-ui/lib/ui/setting/setting_screen.dart
@@ -4,11 +4,11 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:mybrary/res/constants/color.dart';
 import 'package:mybrary/res/constants/config.dart';
 import 'package:mybrary/res/constants/style.dart';
-import 'package:mybrary/ui/common/components/error_page.dart';
 import 'package:mybrary/ui/common/layout/default_layout.dart';
-import 'package:mybrary/ui/common/layout/subpage_layout.dart';
 import 'package:mybrary/ui/profile/profile_edit/profile_edit_screen.dart';
 import 'package:mybrary/ui/setting/components/account_withdrawal.dart';
+import 'package:mybrary/ui/setting/components/login_info.dart';
+import 'package:mybrary/utils/logics/common_utils.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class SettingScreen extends StatelessWidget {
@@ -54,6 +54,16 @@ class SettingScreen extends StatelessWidget {
                         );
                       },
                     ),
+                    _settingTab(
+                      tabTitle: '로그인 정보',
+                      onTap: () {
+                        Navigator.of(context).push(
+                          MaterialPageRoute(
+                            builder: (context) => const LoginInfo(),
+                          ),
+                        );
+                      },
+                    ),
                     const SizedBox(height: 24.0),
                     const Text(
                       '서비스 안내',
@@ -79,14 +89,9 @@ class SettingScreen extends StatelessWidget {
                     _settingTab(
                       tabTitle: '1:1 문의하기',
                       onTap: () async {
-                        String url = inquiryLink;
-                        if (await canLaunchUrl(Uri.parse(url))) {
-                          await launchUrl(
-                            Uri.parse(url),
-                            mode: LaunchMode.externalApplication,
-                            webOnlyWindowName: '_self',
-                          );
-                        }
+                        await connectWebLink(
+                          webLink: inquiryLink,
+                        );
                       },
                     ),
                     const SizedBox(height: 24.0),
@@ -127,13 +132,7 @@ class SettingScreen extends StatelessWidget {
                     const SizedBox(height: 12.0),
                     _settingTab(
                       tabTitle: '리뷰로 응원하기',
-                      onTap: () {
-                        Navigator.of(context).push(
-                          MaterialPageRoute(
-                            builder: (context) => _settingTempPage(),
-                          ),
-                        );
-                      },
+                      onTap: () => connectAppStoreLink(),
                     ),
                     _settingTab(
                       tabTitle: '로그아웃',
@@ -175,22 +174,6 @@ class SettingScreen extends StatelessWidget {
     );
   }
 
-  SubPageLayout _settingTempPage() {
-    return const SubPageLayout(
-      appBarTitle: '리뷰로 응원하기',
-      child: Align(
-        alignment: Alignment.center,
-        child: Column(
-          children: [
-            ErrorPage(
-              errorMessage: '곧 서비스 링크가 열릴 예정이에요!',
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
   Future<dynamic> _showLogoutAlert(
       BuildContext context, FlutterSecureStorage secureStorage) async {
     return await showDialog(
@@ -216,14 +199,14 @@ class SettingScreen extends StatelessWidget {
           actions: [
             Row(
               children: [
-                _confirmButton(
+                confirmButton(
                   onTap: () {
                     Navigator.of(context).pop();
                   },
                   buttonText: '취소',
                   isCancel: true,
                 ),
-                _confirmButton(
+                confirmButton(
                   onTap: () async {
                     await secureStorage.deleteAll();
 
@@ -236,6 +219,8 @@ class SettingScreen extends StatelessWidget {
                   },
                   buttonText: '로그아웃',
                   isCancel: false,
+                  confirmButtonColor: commonRedColor,
+                  confirmButtonText: commonWhiteColor,
                 ),
               ],
             ),
@@ -268,37 +253,6 @@ class SettingScreen extends StatelessWidget {
               'assets/svg/icon/profile_menu_arrow.svg',
             ),
           ],
-        ),
-      ),
-    );
-  }
-
-  Widget _confirmButton({
-    required GestureTapCallback? onTap,
-    required String buttonText,
-    required bool isCancel,
-  }) {
-    return Expanded(
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 4.0),
-        child: InkWell(
-          onTap: onTap,
-          child: Container(
-            height: 46.0,
-            decoration: BoxDecoration(
-              color: isCancel ? greyF1F2F5 : commonRedColor,
-              borderRadius: BorderRadius.circular(4.0),
-            ),
-            child: Center(
-              child: Text(
-                buttonText,
-                style: commonSubBoldStyle.copyWith(
-                  color: isCancel ? commonBlackColor : commonWhiteColor,
-                  fontSize: 14.0,
-                ),
-              ),
-            ),
-          ),
         ),
       ),
     );
